+++ tasks/development/bower.js
@@ -1,12 +1,13 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var bower = lazyReq('bower')();
-var fsExtra = lazyReq('fs-extra')();
-var config = require('../../config');
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const bower = lazyReq('bower');
+const fsExtra = lazyReq('fs-extra');
 
-gulp.task('bower:install', ['bower:copyJson'], (cb) => {
+const config = require('../../config');
 
-  bower.commands.install([], {
+gulp().task('bower:install', ['bower:copyJson'], (cb) => {
+
+  bower().commands.install([], {
     save: true
   }, {})
     .on('end', function (installed) {
@@ -14,8 +15,8 @@ gulp.task('bower:install', ['bower:copyJson'], (cb) => {
     });
 });
 
-gulp.task('bower:copyJson', (cb) => {
-  fsExtra.copySync(
+gulp().task('bower:copyJson', (cb) => {
+  fsExtra().copySync(
     config.bower.jsonFile, 
     './bower.json'
   );