+++ tasks/development/browser-sync.js
@@ -1,11 +1,12 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var browsersync = lazyReq('browser-sync')();
-var config = require('../../config').browsersync.development;
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const browsersync = lazyReq('browser-sync');
+
+const config = require('../../config').browsersync.development;
 
 /**
  * Run the build task and start a server with BrowserSync
  */
-gulp.task('browsersync:development', ['build'], () => {
-  browsersync(config);
+gulp().task('browsersync:development', ['build'], () => {
+  browsersync()(config);
 });