+++ tasks/development/build.js
@@ -1,12 +1,12 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var runSequence = lazyReq('run-sequence')();
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const runSequence = lazyReq('run-sequence');
 
 /**
  * Run all tasks needed for a build in defined order
  */
-gulp.task('build', (cb) => {
-  runSequence(
+gulp().task('build', (cb) => {
+  runSequence()(
     'bower:install',
     'clean',
     'sprites',