+++ tasks/development/clean.js
@@ -1,18 +1,19 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var del = lazyReq('del')();
-var config = require('../../config');
-var cleanOptions = config.clean.options;
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const del = lazyReq('del');
 
-gulp.task('clean', [
+const config = require('../../config');
+const cleanOptions = config.clean.options;
+
+gulp().task('clean', [
   'clean:styles', 
   'clean:scripts:all', 
   'clean:iconFonts'/*,
   'clean:images'*/
 ]);
 
-gulp.task('clean:styles', (cb) => {
-  del([
+gulp().task('clean:styles', (cb) => {
+  del()([
       config.sass.dest + '/**/*.{css,map}'
     ],
     cleanOptions,
@@ -20,8 +21,8 @@ gulp.task('clean:styles', (cb) => {
   );
 });
 
-gulp.task('clean:scripts:all', (cb) => {
-  del([
+gulp().task('clean:scripts:all', (cb) => {
+  del()([
       config.scripts.dest + '**/*.{js,map}',
     ],
     cleanOptions,
@@ -29,8 +30,8 @@ gulp.task('clean:scripts:all', (cb) => {
   );
 });
 
-gulp.task('clean:scripts:head', (cb) => {
-  del([
+gulp().task('clean:scripts:head', (cb) => {
+  del()([
       config.scripts.dest + 'head.min.js',
     ],
     cleanOptions,
@@ -38,8 +39,8 @@ gulp.task('clean:scripts:head', (cb) => {
   );
 });
 
-gulp.task('clean:scripts:vendor', (cb) => {
-  del([
+gulp().task('clean:scripts:vendor', (cb) => {
+  del()([
       config.scripts.dest + 'vendor.js',
     ],
     cleanOptions,
@@ -47,8 +48,8 @@ gulp.task('clean:scripts:vendor', (cb) => {
   );
 });
 
-gulp.task('clean:scripts:app', (cb) => {
-  del([
+gulp().task('clean:scripts:app', (cb) => {
+  del()([
       config.scripts.dest + 'app.{js,map}',
     ],
     cleanOptions,
@@ -56,8 +57,8 @@ gulp.task('clean:scripts:app', (cb) => {
   );
 });
 
-gulp.task('clean:iconFonts', (cb) => {
-  del(config.copy.iconFonts.dest + '*',
+gulp().task('clean:iconFonts', (cb) => {
+  del()(config.copy.iconFonts.dest + '*',
     cleanOptions,
     cb()
   );