+++ tasks/development/jshint.js
@@ -1,14 +1,12 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var jshint = lazyReq('gulp-jshint')();
-var stylish = lazyReq('jshint-stylish')();
-var config = require('../../config').jshint;
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const jshint = lazyReq('gulp-jshint');
+const stylish = lazyReq('jshint-stylish');
 
-/**
- * Check JavaScript sytax with JSHint
- */
-gulp.task('jshint', () => {
-  return gulp.src(config.src)
-    .pipe(jshint())
-    .pipe(jshint.reporter(stylish));
+const config = require('../../config').jshint;
+
+gulp().task('jshint', () => {
+  return gulp().src(config.src)
+    .pipe(jshint()())
+    .pipe(jshint().reporter(stylish()));
 });