+++ tasks/development/sass.js
@@ -1,28 +1,25 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var plumber = lazyReq('gulp-plumber')();
-var browsersync = lazyReq('browser-sync')();
-var sass = lazyReq('gulp-sass')();
-var filter = lazyReq('gulp-filter')();
-var autoprefixer = lazyReq('gulp-autoprefixer')();
-var sourcemaps = lazyReq('gulp-sourcemaps')();
-var config = require('../../config');
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const plumber = lazyReq('gulp-plumber');
+const browsersync = lazyReq('browser-sync');
+const sass = lazyReq('gulp-sass');
+const filter = lazyReq('gulp-filter');
+const autoprefixer = lazyReq('gulp-autoprefixer');
+const sourcemaps = lazyReq('gulp-sourcemaps');
 
-/**
- * Generate CSS from SCSS
- * Build sourcemaps
- */
-gulp.task('sass', () => {
+const config = require('../../config');
+
+gulp().task('sass', () => {
   // Prevent reading sourcemaps to autoprefix them or make sourcemaps of sourcemaps
-  var sassFilter = filter(['*.css', '!*.map']);
+  var sassFilter = filter()(['*.css', '!*.map']);
 
-  return gulp.src(config.sass.src)
-    .pipe(plumber())
-    .pipe(sourcemaps.init())
-    .pipe(sass(config.sass.options).on('error', sass.logError))
-    .pipe(sourcemaps.write('./'))
+  return gulp().src(config.sass.src)
+    .pipe(plumber()())
+    .pipe(sourcemaps().init())
+    .pipe(sass()(config.sass.options).on('error', sass().logError))
+    .pipe(sourcemaps().write('./'))
     .pipe(sassFilter)
-    .pipe(autoprefixer(config.autoprefixer))
+    .pipe(autoprefixer()(config.autoprefixer))
     .pipe(sassFilter.restore())
-    .pipe(gulp.dest(config.sass.dest));
+    .pipe(gulp().dest(config.sass.dest));
 });