+++ tasks/development/sprites.js
@@ -1,40 +1,41 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var spritesmith = lazyReq('gulp.spritesmith')();
-// var buffer = require('vinyl-buffer');
-// var merge = require('merge-stream');
-var config = require('../../config').sprites;
-var fs = lazyReq('fs')();
-var path = lazyReq('path')();
-var glob = lazyReq('glob')().sync;
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const spritesmith = lazyReq('gulp.spritesmith');
+const fs = lazyReq('fs');
+const path = lazyReq('path');
+const glob = lazyReq('glob').sync;
+// const buffer = require('vinyl-buffer');
+// const merge = require('merge-stream');
+
+const config = require('../../config').sprites;
 
 /**
  * Generate sprites and css files from PNGs
  */
-gulp.task('sprites', (cb) => {
+gulp().task('sprites', (cb) => {
   var spritesPath = config.src,
       sprite = {},
       filePath;
 
   // First process base sprite (images not in a subdirectory)
-  sprite.data = gulp.src(path.join(spritesPath, '*.png')).pipe(spritesmith(config.options));
-  sprite.data.img.pipe(gulp.dest(config.dest.image));
-  sprite.data.css.pipe(gulp.dest(config.dest.css));
+  sprite.data = gulp().src(path().join(spritesPath, '*.png')).pipe(spritesmith()(config.options));
+  sprite.data.img.pipe(gulp().dest(config.dest.image));
+  sprite.data.css.pipe(gulp().dest(config.dest.css));
 
-  fs.readdirSync(spritesPath).forEach(function (el, i) {
+  fs().readdirSync(spritesPath).forEach((el, i) => {
     var fileStat,
       spriteName;
-    filePath = path.join(spritesPath, el);
-    fileStat = fs.statSync(filePath);
+    filePath = path().join(spritesPath, el);
+    fileStat = fs().statSync(filePath);
 
     if (fileStat.isDirectory()) {
       spriteName = 'sprite-' + el;
       // Build sprite data from this subdirectory
       config.options.imgName = spriteName + '.png';
       config.options.cssName = spriteName + '.css';
-      sprite.data = gulp.src(path.join(filePath, '*.png')).pipe(spritesmith(config.options));
-      sprite.data.img.pipe(gulp.dest(config.dest.image));
-      sprite.data.css.pipe(gulp.dest(config.dest.css));
+      sprite.data = gulp().src(path().join(filePath, '*.png')).pipe(spritesmith()(config.options));
+      sprite.data.img.pipe(gulp().dest(config.dest.image));
+      sprite.data.css.pipe(gulp().dest(config.dest.css));
     }
   });
   cb();