+++ tasks/production/build.js
@@ -1,12 +1,12 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var runSequence = lazyReq('run-sequence')();
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const runSequence = lazyReq('run-sequence');
 
 /**
  * Run all tasks needed for a build in defined order
  */
-gulp.task('build:production', (cb) => {
-  runSequence(
+gulp().task('build:production', (cb) => {
+  runSequence()(
     'clean:all:production',
     'build', [
       'copy:fonts:production',