+++ tasks/production/clean.js
@@ -1,18 +1,19 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var del = lazyReq('del')();
-var config = require('../../config');
-var cleanOptions = config.clean.options;
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const del = lazyReq('del');
 
-gulp.task('clean:all:production', (cb) => {
-  del(config.clean.production,
+const config = require('../../config');
+const cleanOptions = config.clean.options;
+
+gulp().task('clean:all:production', (cb) => {
+  del()(config.clean.production,
     cleanOptions,
     cb()
   );
 });
 
-gulp.task('clean:scripts:production', (cb) => {
-  del([
+gulp().task('clean:scripts:production', (cb) => {
+  del()([
       config.scripts.productionDest + 'footer.min.{js,map}',
     ],
     cleanOptions,
@@ -20,8 +21,8 @@ gulp.task('clean:scripts:production', (cb) => {
   );
 });
 
-gulp.task('clean:styles:production', (cb) => {
-  del([
+gulp().task('clean:styles:production', (cb) => {
+  del()([
       config.sass.productionDest + '/**/*.min.{css,map}'
     ],
     cleanOptions,