+++ tasks/production/copy.js
@@ -1,13 +1,14 @@
-var lazyReq = require('lazy-req')(require);
-var gulp = lazyReq('gulp')();
-var changed = lazyReq('gulp-changed')();
-var config = require('../../config').copy;
+const lazyReq = require('lazy-req')(require);
+const gulp = lazyReq('gulp');
+const changed = lazyReq('gulp-changed');
+
+const config = require('../../config').copy;
 
 /**
  * Copy fonts to dist folder
  */
-gulp.task('copy:fonts:production', () => {
-  return gulp.src(config.fonts.production.src)
-    .pipe(changed(config.fonts.production.dest))
-    .pipe(gulp.dest(config.fonts.production.dest))
+gulp().task('copy:fonts:production', () => {
+  return gulp().src(config.fonts.production.src)
+    .pipe(changed()(config.fonts.production.dest))
+    .pipe(gulp().dest(config.fonts.production.dest))
 });